fixed bug in shopping cart 'add' procedure
git-svn-id: https://silverstripe-ecommerce.googlecode.com/svn/branches/ssu@1518 950b781c-c544-efab-2f99-90083ee69683

+++ code/control/ShoppingCart.php
@@ -210,7 +210,9 @@ class ShoppingCart extends Object{
 			return false;
 		}
 		if(!$buyable->canPurchase()) {
-			$item->delete();
+			if($item->ID) {
+				$item->delete();
+			}
 			return false;
 		}
 		$quantity = intval($quantity);
